Put back local state on VM to keep edits
If we just rely on upstream change, we have to reformat everything,
so we reintroduce the issues of disallowing invalid intermediate values
and resorting the exclusions while typing. We need local state of edits
and we don't really need to listen to upstream changes atm since
nothing else can change Configuration.