Replace many onX funs with one, taking Intention
Compose docs suggest it's OK to just use the VM in the top-level
Composable, but for now that'd be trickier to create for previews,
and I want to see how this looks a little less coupled. One step at a
time. This is nice for now for just reducing the number of listener funs