Handle the buttons directly in the Composables
Feels a bit weird, but they were in Activity and that was a little too,
so as part of whittling away the code in there, they should either:
A) move into VM and be triggered by an Intention, and they'd then need
to emit something, do that round-trip dance, or have a dependency with
a Context.
B) just be handled in the UI since we already have the Context they
need and never bother the VM or anything else with it.

B seems kinda popular in Compose examples I've seen. Feels a bit
weird, and ofc the presence in the Android docs doesn't guarantee it's
really best, but the implications here are minor anyways so let's
try it out and see how we like it.